+++ bin/informational/disappearing_books
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
 
-from palace.manager.scripts import DisappearingBookReportScript  # noqa: E402
+from palace.manager.do_scripts import DisappearingBookReportScript  # noqa: E402
 
 DisappearingBookReportScript().run()
