+++ bin/util/initialize_instance
@@ -3,7 +3,7 @@
 from pathlib import Path
 
 # Load the alembic configuration relative to this script
-from palace.manager.scripts import InstanceInitializationScript
+from palace.manager.do_scripts import InstanceInitializationScript
 
 alembic_ini = Path(__file__).parent.parent.parent / "alembic.ini"
